Extracts helper methods into helper file
The idea is to split the main spec into several smaller ones; these would share the helper file
